CV07 homework 3 / detection with anchors
Recalculated on GPU with more epochs.
Previously I wrongly assumed I was running it on GPU because Colab GPU runtime was selected. But in fact due to manual `tensorflow` upgrade to **v.2.0.0** (with `pip install`) it was running on CPU because I incorrectly upgraded to the CPU version of `tensorflow` (should've used `tensorflow-gpu` instead).